replace getCurrentClockString lock with atomic
Summary:
Prior to this diff, the current clock tick was maintained
as part of the Synchronized internal view member of InMemoryView.
This made sense because the tick value is incremented while the wlock
on this state is held during observation of the filesystem, and reads
of that value need to be consistent with the writes.

However, it is a little problematic in practice.  In D7075600 we
discovered that calling the `getMostRecentRootNumberAndTickValue`
method (which is a variation on `getCurrentClockString`) in an
error path would block the caller for a surprising amount of time
during a recrawl.

Another problematic aspect impacts the diff I have queued on
top of this one: having these values be part of the internal view
means that we have to dance carefully to avoid recursive locking
concerns.

So, let's move the tick value and the root number (the root number
is a generation counter) outside of the Synchronized container and
replace them with a pair of `std::atomic` values.  This frees the
callers of `getCurrentClockString` and `getMostRecentRootNumberAndTickValue`
from blocking and avoids the potential recursive locking headache.

Reviewed By: simpkins

Differential Revision: D7214110

fbshipit-source-id: ddeab3eece0e802370b9703f3ddd7600e5f46f4e