defer reading of symlink targets
Summary:
I've seen a few problem reports where cold cache situations on laptops
take a very long time to perform an initial crawl.  This diff changes the
recording of symlink targets so that it is no performed eagerly and instead
changes things to use the LRU cache infrastructure that we added for content
hash caching.

Another nice side benefit of this diff is that we save some memory (1x
`w_string` instance per symlink).

Reviewed By: simpkins

Differential Revision: D7341642

fbshipit-source-id: 6c632da6b557667baece71837883567d7cd5abbf