handle scm aware query errors during subscriptions
Summary:
This is likely what was the cause of the blocked client_thread
in D7363065.  I saw this in our coredumper data set:

```
2018-03-22T14:30:20,932: [client=0x7f7b1fd41fd0:stm=0x7f7b1fc42b70:pid=3414331] std::terminate was called. Exception: failed query for the merge base; command returned with status 65280, output= error=.
```

```
(gdb) bt
    at watchman/Logging.h:52
    at ../../.././libstdc++-v3/libsupc++/eh_personality.cc:676
```

from the blacbox log of the user:

```
2018/03/22 14:30:20 user 91113b24fcfd300702f12c81d6ae6738b84313fd (3236283)> killed!
 exited -1 after 2.57 seconds
2018/03/22 14:30:25 user 91113b24fcfd300702f12c81d6ae6738b84313fd (3236618)> killed!
 exited -1 after 6.22 seconds
2018/03/22 14:30:30 user 91113b24fcfd300702f12c81d6ae6738b84313fd (3101237)> killed!
 exited -1 after 2673.23 seconds
```

I suspected that something may have oomed (this is on a devvm), but dmesg
doesn't show any ooming around that time.  The triple kill is an artifact
of the way that chg handles fatal errors (any one child will kill all others
if something breaks), so we need to isolate ourselves from that.

We also need to handle this class of error in a reasonable way.  The
approach here is to swallow the error and continue.  In this code path we
update the query position based on the result of running the query, so if we
get an exception we won't move its cursor forward.  That sounds "ok" to me; the
next time something in the filesystem changes we will effectively retry that
query and perhaps it will succeed.  I don't think there is another reasonable
alternative short of cancelling the subscription, which is a pretty heavy blow.

Reviewed By: simpkins

Differential Revision: D7373808

fbshipit-source-id: c576c3038d5483daf1f31d6f42c3ea311d2e975a