fixup pybser fallback
Summary:
During cmakification I ran into an issue where, if
the `bser.so` python extension had not been properly built, we'd
fall back to the pure python implementation (which is intended behavior)
and then trip over a unicode related error (which is unintended behavior).

In `pybser.py` we were assuming that the parameters had certain
types based on whether we thought we were python 3 or not.  Those
assumptions were not always true and the resolution is to test
that the type matches those expectations.

I've amended the bser tests to strongly verify that the build
successfully generated `bser.so` because this was confusing behavior.

Reviewed By: simpkins

Differential Revision: D7366199

fbshipit-source-id: 0638d24466ccfb611d1c54a9880dd4383c0b122c