fixup internal version reporting
Summary:
This got broken along the way, most likely by some
TARGETS file refactoring.  Since we're moving to cmake for
the open source build, we can simplify the TARGETS file a bit
too.

This removes the probing stuff that sort of helped with building
with opensource buck, and enables the preprocessor define that
pulls in the fbcode build information.

Reviewed By: simpkins

Differential Revision: D7368121

fbshipit-source-id: b5afec3a2dbd70ec9d481a53df0892bd4b978654