add cmake build
Summary:
This is a reasonably functional cmake implementation
of the watchman build system.  Notably missing is support for
detecting pcre and linking it in; that can be done in a follow-up
diff.

This replaces the hand-rolled windows build bits completely,
but does not remove the autoconf bits.  I will likely remove
autoconf once we have parity.

The intent is that this is a stepping stone to adding a dependency
on folly.

When testing the windows build, I found that watchman was missing most of its
registered commands.

It appears as though the relatively tricky constructor attributes that we're
using (which were targeting the `C` compiler) no longer stick; the symbols are
not included in the build and the commands are missing.

The resolution is to use C++ constructor syntax for these instead.

I also tried this on Linux and Mac because it is appealing to
simplify everything, but various versions of gcc/clang on the OSS
CI failed to link in all the commands.

I've adjusted the configuration for the various CI systems so that
we run both autoconf and cmake based builds and tests.  In the longer
run I will remove the autoconf version of the build.

Reviewed By: simpkins

Differential Revision: D7363633

Pulled By: wez

fbshipit-source-id: ac00ae0cd7af5bc51cdcf72c0595c280a4a3f977