fixup compilation error on macos
Summary:
This only seems to be breaking our autoconf based builds
on the mac hosts.  The cmake based builds on macs are succeeding.
Weird!

I've removed the use of container_of; all instances can be
replaced with a simple cast as the members were all at
the front of their respective structs.

Longer term, I'd like to migrate this to folly::dynamic and the
bser implementation there.

```
bser.cpp:345:17: error: offset of on non-standard-layout type 'json_object_t' [-Werror,-Winvalid-offsetof]
  auto object = json_to_object(obj);
```

Reviewed By: chadaustin

Differential Revision: D7849156

fbshipit-source-id: 9d8115a1fc73ee6360d3857644bc736e89205b67