Avoid a race condition about incremental scm-aware queries
Summary:
This is subtle. The problematic case happens in this order:

```
  - [App X] Query watchman with
    {"since": {"scm": {"mergebase-with": B}}}.
    - [watchman] Mergebase is newly seen.
      Call scm->getFilesChangedSinceMergeBaseWith.
      - [hg] "hg status" runs.
        -  [hg] fsmonitor updates its watchman clock to "c:1".
    - [watchman] Get a clock, ex. "c:2". It is returned to App X.
  - [App X] Query watchman with
    {"since": {"scm": {"mergebase-with": B}, "clock": "c:2"}}.
    Intended to get incremental updates since the last query.
    - [watchman] Mergebase does not change.
      Run a pure watchman query since "c:2", without going through scm.
  - Changes between "c:1" and "c:2" might be lost.
```

The issue here is `scm->getFilesChangedSinceMergeBaseWith` has immediate side
effect calling `hg status`, which is affected by the racy working copy. This
patch avoids the issue by moving `getFilesChangedSinceMergeBaseWith` to the
generator so it's deferred.

Note: `scm->mergeBaseWith` is in theory also racy if `.` (working parent)
changes. But that's less a concern - watchman still returns explicit merge
base commit hash with a possibly longer list of files. Things get slower,
but should still be correct.

Reviewed By: wez

Differential Revision: D8081248

fbshipit-source-id: 56a210fae2b3604d8ad3f8b8f74805663ae05d0a