connect cookie change events; fixes state-enter
Summary:
The state-(enter|leave) machinery relies on the observation
of the associated file creation events via the notification mechanism
to correctly sequence state assertions.

Eden mostly doesn't care about this, so the state stuff would never
trigger.

This diff addresses this by having the subscriber thread in the eden
watcher pass the set of created files through to the CookieSync class
to process and wakeup.

Reviewed By: bolinfest

Differential Revision: D8155380

fbshipit-source-id: a72713ff17104e8103017131e19f45cdf8e2a075