batch getSHA1 calls to eden
Summary:
This makes a big difference in the performance of queries that
return the `content.sha1hex` of a large number of files in the warm cache case.

I've taken the simplistic-ish approach of requesting the sha1s after calling
`getFileInformation`.  This could potential run concurrently but may end up
duplicating work if we do that.

Reviewed By: chadaustin

Differential Revision: D8274780

fbshipit-source-id: b63b2cbaf6f657c78721edfe2c7ec6b938a5db90