query expr evaluate returns Optional<bool>
Summary:
This allows us to introduce the concept of a lazy
evaluation of the expression term.   evaluate can now
return `Match`, `NoMatch` or `NeedMoreData` by returning
true, false or nullopt.

The latter is intended to be returned when the expression cannot be evaluated
fully immediately and is somewhat similar to the concept of an `EAGAIN` return
value.

Nothing returns the `NeedMoreData` disposition today but the intention is to
introduce a similar concept to the accessors in the `FileResult` type.

When used in queries like this:

```
["anyof", ["allof", ["type", "f"], ["dirname", "html"], ["anyof", ["suffix", "png"]]]]
```

the lazy evaluation will allow us to avoid fetching full file information
to evaluate the `["type", "f"]` portion if we get a `NoMatch` for the name
based checks later/deeper in the expression tree.

Reviewed By: strager

Differential Revision: D8584085

fbshipit-source-id: 476ba117f08dcb900e8f1219c7658f6cd84b3305