expand lazy fetching of data into result rendering
Summary:
This diff allows the render phase of query processing to return
`Optional<json_ref>` to indicate that it needs data to be loaded before it can
render a given field.

As part of this diff, I've eliminated the `watchman_rule_match` container
struct so that we can reuse the existing `batchFetchData` method.  This has the
nice side effect of reducing the transient memory usage because it can
eliminate a couple of fields that can be derived from the query context.

While that wasn't strictly required for code reuse, a more compelling reason
to remove watchman_rule_match is that it forces eager evaluation of the `new`
flag even when that isn't a requested output field.  Computing the newness
requires a lot of metadata to be loaded, so it is desirable to avoid this
where possible.

This comes at the potential cost of repeating the calculation of the `relname`
(which is the name of the file with any relative root prefix stripped) for some
queries.   There's some more work I want to do to make data fetching lazy, so
I'll revisit the performance of this once I've gotten all the plumbing done.

On a similar note, the logic for rendering futures has a little bit of tension
with batching.  These are only used for content SHA1 and symlink targets today.
I'll make another pass over that aspect of the interface a little later in
this diff stack.

Differential Revision: D8584083

fbshipit-source-id: 235a89fe11df07ef87097939a73afc0381c4a1a7