change FileResult::readLink to non-blocking semantics
Summary:
Prior to this diff, the readLink call returned a Future to allow
asynchronously producing the readLink results.  This has is in tension with the
infrastructure to batch process render results, so this diff makes it follow
similar semantics to the other methods of FileResult; if it returns an empty
optional then we will need to load the target later during a batch fetch
operation.

There are currently no implementations of this class that have an actual batch
implementation of this function, so this is slightly moot, however, it is
beneficial because it is a step to removing the Future flavor of these
accessors to relieve the tension mentioned above.

While there are no batch implementations of readLink, the InMemoryView implementation
can run the underlying Futures for the cache lookups concurrently.

Bugs discovered:

* LocalFileResult wasn't respecting the semantics of returning a nullptr w_string() for !symlink files.
* EdenFileResult hadn't hooked up an implementation of readLink.  We do the dumb thing here.

Reviewed By: strager

Differential Revision: D8584088

fbshipit-source-id: db3a98a3f61e49429c993a1d99cfb4aa899ab02e