pre-set exists status for fresh instance queries in eden
Summary:
When we've produced a file list by evaluating a glob, we know
that the files exist, so let's remember that.  This helps to avoid
querying FullFileInformation for expressions that use the `exists` term;
this can save querying a significant volume of metadata for fresh
instance results in large repos.

Reviewed By: strager

Differential Revision: D8584081

fbshipit-source-id: eb925fab2dbf2da69603a5a2b4cfe5584df6f4b3