enable hg fsmonitor race detection during `hg status`
Summary:
This diff enables the new fsmonitor config option that
requests strict race detection during `hg status` operations.
When this config is enabled, mercurial will raise an error
if at the end of a status call it detects that the working
copy has been changed in a way that will produce misleading
or bad results rather than being oblivious to them.

This config option is safe to enable even if the mercurial
install doesn't have support for this functionality; it will
be silently ignored.

When the flag is enabled and an issue is detected, the status
call will error out and bubble up an error to the watchman
query executor.  For `query` commands this will yield an `error`
PDU.

For subscriptions it will cause the current batch to error
out, but because mutation happened or is happening, the subscription
result processing is likely to trigger again almomst instantaneously
and emit the correct current results.

We had talked about potentially making the `query` path retry the
call in the error case, but it's not clear that the complexity in
handling the retry is warranted, so let's try this to start.

Reviewed By: quark-zju

Differential Revision: D8848628

fbshipit-source-id: 699646200d299e91c1f558dba861ed0d2bfc4f5a