Add a local saved state implementation to Watchman
Summary:
Add a local saved state manager implementation that finds saved
states stored in the local file system. Use the most recent n revisions
to find the saved state, and return a JSON error message if none can
be found.

Throw if a saved state spec is specified in the clockspec, yet it is
missing the required local storage path or project, or has the wrong type
for any of its parameters including the optional max revs (defaults to
10, but could happily remove the default if desired).

This is primarily for integration testing the interface and the actual
integration into Watchman queries/subscriptions, but it should be
functional if something were to write these states locally.

Reviewed By: wez

Differential Revision: D8263617

fbshipit-source-id: fbe3bf719ee1c11e72ca073c45467031122f821f