Add XDB client for saved state XDB
Summary:
Pull the XDB-specific parts of the manifold saved state manager
out into a separate class. This class should do nothing but provide
an interface to the xdb. The only method it provides returns the most
recent saved state revision with a commit date less than or equal to
the given source control commit date.

Reviewed By: wez

Differential Revision: D8292011

fbshipit-source-id: 9bfcb9f01a998a2e27b6621813943bf7176fc0c4