Add manifold client for saved state
Summary:
Pull manifold code out into its own class that contains only the
code for accessing manifold. Provide a method to get the properties for
a revision that's present in manifold, and a method to get the manifold
path, which we'll return to the client so they can access the actual data.

Eventually we'll likely want to access the actual data here for
prefetching, but that's for a later time.

On failure, log the error but just return a null pointer, because we do
not want failures to load saved state info to be blocking.

Pull the manifold bucket, tier, and timeout settings from the watchman
configuration if specified, but fall back to reasonable defaults.

Reviewed By: wez

Differential Revision: D8295319

fbshipit-source-id: d463d55ad6ef0b44dc0ebad11ded1d079f2335bc