Add a manifold watchman saved state implementation
Summary:
Add a facebook-specific manifold saved state implementation that
checks for the most recent saved state based on timestamp from a saved
state xdb, then loads the properties for that saved state from manifold
and returns both the properties and manifold path so that clients can
access the saved state itself via the path.

The properties include info like the hack server version, hack config,
FactsDB version, etc., that may indicate to the client whether or not
the saved state is useful.

As a follow up, I plan to prefetch the saved state and include info
regarding how to access the prefetched saved state. But, that will come
in a later diff.

Reviewed By: wez

Differential Revision: D8267113

fbshipit-source-id: cb3c717b2bc9dfdca190531a29696f3510d1e8c3