fix backtrace.cpp compilation on vs 2017
Summary:
A combination of the newer compiler and the headers pulled
in by folly portability cause compilation failures in this module
because nothing has defined the `IN`, `OUT`, or `OPTIONAL` parameter
markers.

This diff defines them to the empty string if they haven't already
been set up.

I couldn't find a single reasonable header file to include to
satisfy this requirement that didn't cause other compilation issues,
so this is the minimal pragmatic approach.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9396872

fbshipit-source-id: 04b2bc143591f716b72bff4ebfba519b9878faa2