prevent accidental spawns in test suite
Summary:
A confluence of issues could lead to client(s) accidentally
spawning a non-isolated watchman instance during the integration test run.
The two main issues that contribute are:

* Misinterpretation of connection errors
* Race conditions with named pipe creation

In these circumstances, the client may choose to run `watchman get-sockname`
as part of a fallback.  When this happens, none of the settings from the
isolated environment of the test suite are propagated through to the
watchman executable when it wants to spawn the server.

This diff adds a kill switch of sorts that we can set at the start of
the tests; when the environment variable is set, it implicitly sets
the `--no-spawn` command line parameter.   This prevents "breaking out"
of the test environment.

Refs: https://github.com/facebook/watchman/pull/636

Reviewed By: simpkins

Differential Revision: D9396870

fbshipit-source-id: 54ee8e9cd6b83fa9b9dd6319f8eec91dc5f5868d